Fixed relation from input to gray code, adjust to detent position.
Adjusted the decoding table, so that it is indexed by counter values
instead of "old_quadrature_input". We can get rid of one varibale
that way, also we have a fixed relation from input states to the
2 MSBs of the counter.

As the counters are now directly related to the quadrature inputs,
we can further adjust the point at which the output signals change
to one particular state change in quadrature input signals. Hence
we can make the counter display change more closely to the
physical detent, this improves the "feel" when changing parameters.